move utils to shared
