+++ src/shared/utils/cloudinary.js
